feat: 404에러 핸들링
전역적으로 없는 아이디 조회시 404에러를 반환.

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,6 +1,8 @@
 package roomescape.reservation.service;
 
 import org.springframework.stereotype.Service;
+import roomescape.common.exception.NotFoundException;
+import roomescape.common.exception.ExceptionMessage;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.dto.ReservationRequestDto;
 import roomescape.reservation.repository.ReservationRepository;
@@ -21,8 +23,8 @@ public class ReservationService {
 
     public Reservation save(ReservationRequestDto reservationRequestDto) {
         ReservationTime time = reservationTimeRepository.findByStartAt(reservationRequestDto.time());
-        if(time == null) {
-            throw new IllegalArgumentException("Time not found");
+        if (time == null) {
+            throw new NotFoundException(ExceptionMessage.RESERVATION_TIME_NOT_FOUND_BY_TIME.getMessage() + reservationRequestDto.time());
         }
 
         Reservation requestReservation = reservationRequestDto.toEntity(time);
@@ -42,7 +44,10 @@ public class ReservationService {
     }
 
     public void deleteById(Long id) {
-        //no content 예외처리
+        if (reservationRepository.findById(id) == null) {
+            throw new NotFoundException(ExceptionMessage.RESERVATION_NOT_FOUND_BY_ID.getMessage() + id);
+        }
+
         reservationRepository.deleteById(id);
     }
 }
